backend/requirements.txt
# FastAPI 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 데이터베이스 ORM
sqlalchemy==2.0.23

# JWT 인증
python-jose[cryptography]==3.3.0

# HTTP 클라이언트
httpx==0.25.2
requests==2.31.0

# 환경변수 관리
python-dotenv==1.0.0

# 데이터 검증 및 직렬화
pydantic==2.5.0

# 파일 업로드 처리
python-multipart==0.0.6

# Cloudflare API
cloudflare==2.19.4

# 비동기 지원
asyncio-mqtt==0.16.1

# 날짜/시간 처리
python-dateutil==2.8.2
